Adding JavaDoc to: entities, services, controller etc.

+++ MunicipalPlatform-Java/MunicipalPlatform/MunicipalPlatform/src/main/java/it/unicam/cs/idsflsm/municipalplatform/application/services/user/UserService.java
@@ -1,18 +1,17 @@
 package it.unicam.cs.idsflsm.municipalplatform.application.services.user;
-import it.unicam.cs.idsflsm.municipalplatform.application.commands.user.IRoleCommand;
-import it.unicam.cs.idsflsm.municipalplatform.application.commands.user.RemoveRoleCommand;
-import it.unicam.cs.idsflsm.municipalplatform.application.commands.user.UpdateRoleCommand;
+import it.unicam.cs.idsflsm.municipalplatform.application.commands.user.authenticated.RemoveRoleCommand;
+import it.unicam.cs.idsflsm.municipalplatform.application.commands.user.authenticated.UpdateRoleCommand;
 import it.unicam.cs.idsflsm.municipalplatform.application.configurations.PasswordEncoder;
 import it.unicam.cs.idsflsm.municipalplatform.application.abstractions.services.user.IUserService;
-import it.unicam.cs.idsflsm.municipalplatform.application.factories.user.AuthenticatedUserBuilderFactory;
+import it.unicam.cs.idsflsm.municipalplatform.application.factories.user.authenticated.AuthenticatedUserBuilderFactory;
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.user.RoleRequestMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.user.anonymous.TouristMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.user.authenticated.GenericAuthenticatedUserMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.user.RoleRequestDto;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.user.anonymous.AnonymousUserDto;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.user.authenticated.AuthenticatedUserDto;
-import it.unicam.cs.idsflsm.municipalplatform.application.models.requests.user.LoginRequest;
-import it.unicam.cs.idsflsm.municipalplatform.application.models.requests.user.RegisterRequest;
+import it.unicam.cs.idsflsm.municipalplatform.application.models.requests.user.authenticated.LoginRequest;
+import it.unicam.cs.idsflsm.municipalplatform.application.models.requests.user.authenticated.RegisterRequest;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.user.RoleRequest;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.user.anonymous.Tourist;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.user.authenticated.AuthenticatedUser;
@@ -26,17 +25,34 @@ import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.Predicate;
-
+/**
+ * Service class for the AuthenticatedUserRepository, AnonymousUserRepository and RoleRequestRepository.
+ * It provides methods to manipulate persistent authenticated and anonymous users, and their requested roles,
+ * in the database
+ */
 @Service
 @Transactional
 @AllArgsConstructor(onConstructor_ = @Autowired)
 public class UserService implements IUserService {
+    /**
+     * The repository for AuthenticatedUser entity
+     */
     private final IAuthenticatedUserRepository _authenticatedUserRepository;
+    /**
+     * The repository for AnonymousUser entity
+     */
     private final IAnonymousUserRepository _anonymousUserRepository;
-    private final IRoleRequestRepository _requestRoleRepository;
+    /**
+     * The repository for RoleRequest entity
+     */
+    private final IRoleRequestRepository _roleRequestRepository;
+    /**
+     * The password encoder for encoding user passwords
+     */
     private final PasswordEncoder _passwordEncoder;
     @Override
     public void saveInRepository(AuthenticatedUser authenticatedUser) {
@@ -46,66 +62,63 @@ public class UserService implements IUserService {
     public AuthenticatedUserDto register(RegisterRequest request) {
         var existingUser = findByUsername(request.getEmail());
         if (existingUser == null) {
-            // TODO : DO NOT FORGET
             AuthenticatedUserBuilderFactory factory = new AuthenticatedUserBuilderFactory();
-            var builder = factory.createAuthenticatedUserBuilder(UserRole.ANIMATOR);
+            var builder = factory.createAuthenticatedUserBuilder(UserRole.AUTHENTICATED_TOURIST);
             builder.setUsername(request.getEmail());
             builder.setPassword(_passwordEncoder.encodePassword(request.getPassword()));
             builder.setName(request.getName());
             builder.setSurname(request.getSurname());
-            builder.setRole(UserRole.ANIMATOR);
-//            user.setRole(UserRole.AUTHENTICATED_TOURIST);
+            builder.setRole(UserRole.AUTHENTICATED_TOURIST);
             var user = builder.build();
             user.setId(UUID.randomUUID());
             user = _authenticatedUserRepository.save(user);
             return GenericAuthenticatedUserMapper.toDto(user, true);
-        } else {
-            return null;
         }
+        return null;
     }
     @Override
     public AuthenticatedUserDto login(LoginRequest request) {
-        var existingUser = _authenticatedUserRepository.findByUsername(request.getEmail());
-        if (existingUser != null) {
-            if (_passwordEncoder.verifyPassword(request.getPassword(), existingUser.getPassword())) {
-                return GenericAuthenticatedUserMapper.toDto(existingUser, true);
-            } else {
-                return null;
+        List<AuthenticatedUser> users = _authenticatedUserRepository.findByUsername(request.getEmail());
+        if (!users.isEmpty()) {
+            AuthenticatedUser user = users.get(0);
+            if (_passwordEncoder.verifyPassword(request.getPassword(), user.getPassword())) {
+                return GenericAuthenticatedUserMapper.toDto(user, true);
             }
-        } else {
             return null;
         }
+        return null;
     }
     @Override
     public RoleRequestDto sendRequestedRole(RoleRequestDto requestDto) {
-        if (_authenticatedUserRepository.findByUsername(requestDto.getUsername()) != null) {
+        if (!_authenticatedUserRepository.findByUsername(requestDto.getUsername()).isEmpty()) {
             RoleRequest roleRequest = RoleRequestMapper.toEntity(requestDto);
-            _requestRoleRepository.save(roleRequest);
+            _roleRequestRepository.save(roleRequest);
             return RoleRequestMapper.toDto(roleRequest);
-        } else {
-            return null;
         }
+        return null;
     }
     @Override
     public List<RoleRequestDto> receiveRequestedRoles(Predicate<RoleRequest> predicate) {
-        return RoleRequestMapper.toDto(_requestRoleRepository.findAll().stream().filter(predicate).toList());
+        List<RoleRequest> roleRequests = _roleRequestRepository.findAll().stream().filter(predicate).toList();
+        if (!roleRequests.isEmpty()) {
+            return RoleRequestMapper.toDto(roleRequests);
+        }
+        return new ArrayList<>();
     }
     @Override
     public AuthenticatedUserDto updateUserRole(UUID idUser, boolean accept) {
         AuthenticatedUser authenticatedUser = _authenticatedUserRepository.findById(idUser).orElse(null);
-        RoleRequest roleRequest = (authenticatedUser != null) ? _requestRoleRepository.findById(authenticatedUser.getUsername()).orElse(null) : null;
+        RoleRequest roleRequest = (authenticatedUser != null) ? _roleRequestRepository.findById(authenticatedUser.getUsername()).orElse(null) : null;
         if (authenticatedUser != null && roleRequest != null) {
             if (accept) {
-                // authenticatedUser.setCommand(new UpdateRoleCommand(authenticatedUser, roleRequest.getRole()));
-                var newUser = swapRoles(authenticatedUser, roleRequest);
                 _authenticatedUserRepository.delete(authenticatedUser);
+                var newUser = swapRoles(authenticatedUser, roleRequest);
                 _authenticatedUserRepository.save(newUser);
             }
-            _requestRoleRepository.delete(roleRequest);
+            _roleRequestRepository.delete(roleRequest);
             return GenericAuthenticatedUserMapper.toDto(authenticatedUser, true);
-        } else {
-            return null;
         }
+        return null;
     }
     private AuthenticatedUser swapRoles(AuthenticatedUser authenticatedUser, RoleRequest roleRequest) {
         AuthenticatedUserBuilderFactory factory = new AuthenticatedUserBuilderFactory();
@@ -118,8 +131,6 @@ public class UserService implements IUserService {
         builder.setSurname(authenticatedUser.getSurname());
         builder.setPois(authenticatedUser.getPois());
         builder.setItineraries(authenticatedUser.getItineraries());
-//         TODO : CHEATED HERE!
-        // builder.setRole(authenticatedUser.getRole());
         var user = builder.build();
         user.setId(authenticatedUser.getId());
         if (roleRequest != null) {
@@ -133,7 +144,7 @@ public class UserService implements IUserService {
     @Override
     public AuthenticatedUserDto removeUserRole(UUID idUser) {
         AuthenticatedUser authenticatedUser = _authenticatedUserRepository.findById(idUser).orElse(null);
-        if (authenticatedUser != null) {
+        if (authenticatedUser != null && !authenticatedUser.getRole().equals(UserRole.ADMINISTRATOR)) {
             authenticatedUser.setCommand(new RemoveRoleCommand(authenticatedUser));
             var newUser = swapRoles(authenticatedUser, null);
             _authenticatedUserRepository.delete(authenticatedUser);
@@ -142,14 +153,6 @@ public class UserService implements IUserService {
         } else {
             return null;
         }
-    }
-    @Override
-    public RoleRequestDto findRoleRequestByUsername(String username) {
-        return RoleRequestMapper.toDto(_requestRoleRepository.findAll()
-                .stream()
-                .filter(roleRequest -> roleRequest.getUsername().equals(username))
-                .findFirst()
-                .orElse(null));
     }
     @Override
     public AuthenticatedUserDto findByUsername(String username) {
@@ -174,27 +177,27 @@ public class UserService implements IUserService {
             return authenticatedUser.getRole().getPermissions().contains(permission);
         } else {
             var anonymousUser = _anonymousUserRepository.findById(idUser).orElse(null);
-            return anonymousUser != null;
+            if (anonymousUser != null) {
+                return anonymousUser.getRole().getPermissions().contains(permission);
+            }
+            return false;
         }
     }
-
     @Override
     public AnonymousUserDto openAnonymousSession() {
         Tourist tourist = new Tourist(UUID.randomUUID());
         if (!_anonymousUserRepository.existsById(tourist.getId())) {
             _anonymousUserRepository.save(tourist);
             return TouristMapper.toDto(tourist);
-        } else {
-            return null;
         }
+        return null;
     }
     @Override
     public boolean closeAnonymousSession(AnonymousUserDto userDto) {
         if (_anonymousUserRepository.existsById(userDto.getId())) {
             _anonymousUserRepository.deleteById(userDto.getId());
             return true;
-        } else {
-            return false;
         }
+        return false;
     }
 }
